Moved announcements components to Toast component as that's the only place they should be used
